+++ litemapy/__init__.py
@@ -1,2 +1,6 @@
-from .schematic import Schematic, Region, BlockState, Entity, TileEntity
+from .Schematic import Schematic
+from .Region import Region
+from .BlockState import BlockState
+from .Entity import Entity
+from.TileEntity import TileEntity
 from .info import LITEMAPY_VERSION as __version__
